Replace spraypaint with swr
The swr library is the Vercel recommended way of fetching data in nextjs
apps and integrates with the React Hooks API which removes a lot of the
boilerplate code. Since it's not possible to get the .NET API to return
JSON:API compliant responses continuing with spraypaint makes no sense.

This also tweaks the operative summary to use the new object for trade.