Use explicit non-productive flag for non-product types
We're likely to add additional pay element types for overtime and
out-of-hours so rather than rely on the implicit nature of not being
the other types add an explicit flag for non-productive types so that
we don't need to change the code every time we add a new flag.

Depends on LBHackney-IT/bonuscalc-api#41 being be merged and deployed.