Show past weeks instead of closed weeks
Complaints were made about no data being in reports because weeks weren't
closed so this is an option to use pasts weeks instead of closed weeks.

The downside to this is that if reports are sent out when weeks are
closed then it will include data from later weeks - this may or may not
be a problem as it could alert operatives to issues in their timesheets.