Filter out archived operatives when closing a week
When an operative has been archived they will always have zero SMVs for a
week until the end of the bonus period at which point they're not added to
the next period. Since this is confusing and time-wasting since it has to
be checked then filter them out specifically during the closing of the week.